+++ pgpack/__init__.py
@@ -23,4 +23,4 @@ __all__ = (
     "PGPackWriter",
 )
 __author__ = "0xMihalich"
-__version__ = "0.1.3.1"
+__version__ = "0.2.0.0"